corrección de errores

+++ PI_2018/content/activity_manager.php
@@ -1,21 +1,18 @@
 <?php
-
 session_start();
 include '../php/connection.php';
-
-
 ?>
 <!DOCTYPE html>
 <html>
 <head>
-	<link rel="stylesheet" type="text/css" href="../css/style.css">
-	<meta charset="utf-8">
-	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-  	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-  	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-  	<script type="text/javascript" src="../js/conectores.js"></script>
-	<script type="text/javascript" src="../js/main.js"></script>
-	<title>Administrar ofertas</title>
+  <link rel="stylesheet" type="text/css" href="../css/style.css">
+  <meta charset="utf-8">
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+    <script type="text/javascript" src="../js/conectores.js"></script>
+  <script type="text/javascript" src="../js/main.js"></script>
+  <title>Administrar ofertas</title>
 </head>
 <body>
 <!-- ESTO ES LO MISMO QUE EL COMPONENTE, BASTA CON VOLVER A SUSTITUIRLO, CON LA DIFERENCIA DE QUE EL COMPONENTE NO ENLAZA BIEN CON EL LOGIN -->
@@ -25,33 +22,29 @@ include '../php/connection.php';
       <a class="navbar-brand" href="#">Nombre Web</a>
     </div>
     <ul class="nav navbar-nav navbar-right">
-    	<?php
+      <?php
         /**
-         *      Submenú lateral de usuario/empresa conectado
+         *
+         *      Sesión del usuario (no empresa)
+         *
          */
-
-        // Si el usuario NO está logueado.
-    		if (!isset($_SESSION['nombre'])) {
-    			echo '<li><a href="content/registrouser.html"><span class="glyphicon glyphicon-download-alt"></span> Registrarse</a></li>';
-      			echo '<li><a href="content/login.html"><span class="glyphicon glyphicon-log-in"></span> Entrar</a></li>';
-        // Si el usuario está logueado, y visualizará distintos menús dependiendo de si es EMPRESA o USUARIO.
-    		}else{
-      			echo '
+            //$nombreuser = $_SESSION['nombre']; // 'Alias' del usuario que ha iniciado sesión, da error cuando no está iniciada la sesión porque dicha variable queda vacía.
+        if (!isset($_SESSION['nombre'])) {
+          echo '<li><a href="content/registrouser.html"><span class="glyphicon glyphicon-download-alt"></span> Registrarse</a></li>';
+            echo '<li><a href="content/form_login.html"><span class="glyphicon glyphicon-log-in"></span> Entrar</a></li>';
+        }else{
+            echo '
                 <li class="dropdown">
                     <a class="dropdown-toggle" data-toggle="dropdown" href="#"><span class="glyphicon glyphicon-user"></span> ' . $_SESSION['nombre']  . '
                     <span class="caret"></span></a>
-                    <ul class="dropdown-menu">';
-      			if ($_SESSION['tipo'] === "usuario") {
-      			    echo '<li><a href="perfil.php"><span class="glyphicon glyphicon-cog"></span> Mi perfil</a></li>
-                          <li><a href="ofertas.php"><span class="glyphicon glyphicon-th-list"></span> Mis reservas</a></li>';
-                } else if ($_SESSION['tipo'] === "empresa") {
-                    echo '<li><a href="perfil.php"><span class="glyphicon glyphicon-cog"></span> Perfil de empresa</a></li>';
-                }
-      			echo '<li><a href="php/logout.php"><span class="glyphicon glyphicon-log-out"></span> Cerrar sesión</a></a></li>
+                    <ul class="dropdown-menu">
+                      <li><a href="perfil.php"><span class="glyphicon glyphicon-cog"></span> Mi perfil</a></li>
+                      <li><a href="ofertas.php"><span class="glyphicon glyphicon-th-list"></span> Mis reservas</a></li>
+                      <li><a href="php/logout.php"><span class="glyphicon glyphicon-log-out"></span> Cerrar sesión</a></a></li>
                     </ul>
                 </li>';
-    		}
-    	?>
+        }
+      ?>
     </ul>
   </div>
 </header>
@@ -65,53 +58,52 @@ include '../php/connection.php';
 <section>
 <article>
 <?php
-
-      /* Se encarga de borrar el registro seleccionado pulsando el botón que tenemos en la tabla. 
-
-        Importante el saber que cuando tenemos reservas de dicha actividad esta no se puede borrar, ya que el id de la oferta está asociada a la reserva. Esto se controlará en futuras versiones.
+      /* Se encarga de borrar el registro seleccionado pulsando el botón que tenemos en la tabla. Solo se puede borrar una oferta si esta no tiene reservas pendientes. */
+    if(isset($_GET['aski']) == 'delete'){
         
-      */
-
-      if(isset($_GET['aski']) == 'delete'){
-
         $id = mysqli_real_escape_string($conexion,(strip_tags($_GET["id"],ENT_QUOTES)));
-        $sql_del = "DELETE FROM oferta WHERE id='$id'";
-        $delete = $conexion->query($sql_del);
+        $sql_reservas = "SELECT * FROM reserva WHERE id='$id'"; //comprobamos que no haya reservas.
+        $comprobar = $conexion->query($sql_reservas);
+        
+        if ($comprobar->num_rows > 0) {
+            echo '<div class="alert alert-danger alert-dismissable"><a href="ofertas.php"><button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button></a> No se puede borrar el registro porque tiene una reserva asociada.</div>';
+        }else{
+          
+          $sql_del = "DELETE FROM oferta WHERE id='$id'";
+          $delete = $conexion->query($sql_del);
+          
           if($delete){
-            echo '<div class="alert alert-success alert-dismissable"><button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button> Actividad eliminada correctamente.</div>';
+              echo '<div class="alert alert-success alert-dismissable"><button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button> Datos eliminado correctamente.</div>';
           }else{
-            echo '<div class="alert alert-danger alert-dismissable"><button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button> Error, no se pudo eliminar los datos.</div>';
-            echo "Error: " . $sql_del . "<br>" . $conexion->error;
-          }
+              echo '<div class="alert alert-danger alert-dismissable"><button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button> Error, no se pudo eliminar los datos --> ';
+            }
         }
+    }
       
-      ?>
+?>
 </article>
 <article>
     <h2 class="text-center">Mis actividades</h2><br/>
-		<div class="table-responsive">
-			<table class="table table-striped table-hover">
+    <div class="table-responsive">
+      <table class="table table-striped table-hover">
         <thead>
-				<tr>
-					<th>Nº</th>
+        <tr>
+          <th>Nº</th>
           <th>Actividad</th>
           <th>Tipo de actividad</th>
           <th>Localización</th>
-					<th>Dificultad</th>
           <th>Precio</th>
           <th>Inicio de la actividad</th>
           <th>Fin de la actividad</th>
-					<th>Acciones</th>
-				</tr>
+          <th>Dificultad</th>
+          <th>Acciones</th>
+        </tr>
         </thead>
 
-				<?php
-
+        <?php
         /* Este es el data-table que mostrará los datos de las actividades publicadas por la empresa.*/
-
-				$sql_oferta = "SELECT * FROM oferta WHERE cif_empresa = (SELECT cif FROM empresa WHERE alias = '". $_SESSION['nombre']  ."') "; 
-
-				  $result = $conexion->query($sql_oferta); 
+        $sql_oferta = "SELECT * FROM oferta WHERE cif_empresa = (SELECT cif FROM empresa WHERE alias = '". $_SESSION['nombre']  ."') "; 
+          $result = $conexion->query($sql_oferta); 
           $no = 1;
           if ($result->num_rows === 0) {
             echo '<tr><td colspan="8">No hay actividades.</td></tr>';
@@ -123,6 +115,9 @@ include '../php/connection.php';
               <td>'.$row['nombre'].'</td>
               <td>'.$row['tipo_actividad'].'</td>
               <td>'.$row['localizacion'].'</td>
+              <td>'.$row['precio'].' €</td>
+              <td>'.$row['fecha_inicio'].'</td>
+              <td>'.$row['fecha_fin'].'</td>
               <td>';
               if($row['dificultad'] == 'facil'){
                 echo '<span class="label label-success">Fácil</span>';
@@ -137,15 +132,12 @@ include '../php/connection.php';
                 echo '<span class="label label-danger">Experto</span>';
               }
             echo '
-              </td>
-              <td>'.$row['precio'].' €</td>
-              <td>'.$row['fecha_inicio'].'</td>
-              <td>'.$row['fecha_fin'].'</td>
+              </td>              
               <td>
  
                 <a href="actividad.php?id='.$row['id'].'" title="Ver" class="btn btn-success btn-sm"><span class="glyphicon glyphicon-zoom-in" aria-hidden="true"></span></a>
                 <a href="edit.php?id='.$row['id'].'" title="Editar datos" class="btn btn-primary btn-sm"><span class="glyphicon glyphicon-cog" aria-hidden="true"></span></a>                
-                <a href="ofertas.php?aski=delete&id='.$row['id'].'" title="Eliminar" onclick="return confirm(\'¿Está seguro de que quiere anular la reserva?\')" class="btn btn-danger btn-sm"><span class="glyphicon glyphicon-trash" aria-hidden="true"></span></a>
+                <a href="activity_manager.php?aski=delete&id='.$row['id'].'" title="Eliminar" onclick="return confirm(\'¿Está seguro de que quiere anular la reserva?\')" class="btn btn-danger btn-sm"><span class="glyphicon glyphicon-trash" aria-hidden="true"></span></a>
               </td>
             </tr>
             ';
@@ -154,13 +146,13 @@ include '../php/connection.php';
         }
         ?>
           
-			</table>
-			</div>
-		</div>
+      </table>
+      </div>
+    </div>
 
 </article>
 
-	
+  
 </section>
 
 <footer>
